Refactor(kjh) : groupbuying요청사항 수정 #2
