+++ src/main/java/com/example/livingbymyselfserver/chat/entity/dto/ChatMessageResponseDto.java
@@ -3,7 +3,6 @@ package com.example.livingbymyselfserver.chat.entity.dto;
 import com.example.livingbymyselfserver.chat.entity.Chat;
 import lombok.Getter;
 
-
 @Getter
 public class ChatMessageResponseDto {
 
@@ -11,7 +10,6 @@ public class ChatMessageResponseDto {
   private String msg;
   private String localTime;
   private ChatResponseDto responseDto;
-
 
   public ChatMessageResponseDto(Chat chat) {
     this.chatId = chat.getId();
