+++ src/main/java/com/example/livingbymyselfserver/chat/entity/dto/ChatRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.livingbymyselfserver.chat;
+package com.example.livingbymyselfserver.chat.entity.dto;
 
 import lombok.Getter;
